requirements.txt
# Core dependencies
Jinja2>=3.1.0
PyYAML>=6.0
python-docx>=0.8.11

# GUI dependencies (optional)
# tkinter is included with Python standard library
PyQt5>=5.15.0; extra == "pyqt5"

# Development and testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0

# Additional utilities
pathlib2>=2.3.0; python_version < "3.4"
dataclasses>=0.6; python_version < "3.7"